+++ buildSrc/src/main/java/Versions.kt
@@ -56,7 +56,7 @@ object Versions {
     const val coil = "2.6.0"
 
     // Media
-    const val media3 = "1.0.0"
+    const val media3 = "1.3.1"
 
     // Logging
     const val timber = "5.0.1"